+++ java-service-processor/src/test/resources/definition/TestNestedLoaderCustomNames.java
@@ -1,0 +1,14 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+
+public class TestNestedLoaderCustomNames {
+
+    @ServiceDefinition(loaderName = "definition.{{topLevelClassName}}Loader")
+    public interface FooService {
+    }
+
+    @ServiceDefinition(loaderName = "definition.{{topLevelClassName}}Loader")
+    public interface BarService {
+    }
+}